backend/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
psycopg2-binary==2.9.7
pymongo==4.4.0
google-cloud-pubsub==2.19.0
python-jose==3.3.0
requests==2.31.0
livekit-api==0.6.3
pytest==7.4.0
